papers2code_app2/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pymongo
passlib[bcrypt]
requests
python-dotenv
dateutils
slowapi
